+++ changelog.md
@@ -60,3 +60,7 @@
 ## 1.2.1
 
 * Travis runs tests on node 0.10 and 0.12
+
+## Next
+
+* Run jshint on test code